Fix implementation of enhanced double hashing
Use the correct loop counter for advancing the hash function.

Thanks to Amos Wenger for noticing this.

Changing this causes the behaviour to change slightly, which requires
the test to be updated. Due to randomness it's not enough anymore to
estimate the ratio of false positives with 10000 random values but 500000
are required now to get close enough to the theoretical value.

This does not mean that there are more false positives now, just that
the random values being tested are layed out in a way that more tests
are needed to estimate the false positive ratio reliably.